LayoutTests/imported/w3c/web-platform-tests/uievents/constructors/inputevent-constructor.html

<!DOCTYPE html>
<title>InputEvent Constructor Tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var e = new InputEvent('type');
  assert_equals(e.data, null, '.data');
  assert_false(e.isComposing, '.isComposing');
}, 'InputEvent constructor without InputEventInit.');

test(function() {
  var e = new InputEvent('type', { data: null, isComposing: true });
  assert_equals(e.data, null, '.data');
  assert_true(e.isComposing, '.isComposing');
}, 'InputEvent construtor with InputEventInit where data is null');

test(function() {
  assert_equals(new InputEvent('type', { data: ''}).data, '', '.data');
}, 'InputEvent construtor with InputEventInit where data is empty string');

test(function() {
  assert_equals(new InputEvent('type', { data: 'data' }).data, 'data', '.data');
}, 'InputEvent construtor with InputEventInit where data is non empty string');

test(function() {
  var e = new InputEvent('type', { dataTransfer: null });
  assert_equals(e.dataTransfer, null, '.dataTransfer');
}, 'InputEvent constructor with InputEventInit where dataTransfer is null');

test(function() {
  var dataTransfer = new DataTransfer();
  var e = new InputEvent('type', { dataTransfer: dataTransfer });
  assert_equals(e.dataTransfer, dataTransfer, '.dataTransfer');
}, 'InputEvent constructor with InputEventInit where dataTransfer is set');

test(function() {
  var e = new InputEvent('type', { targetRanges: [] });
  assert_equals(e.getTargetRanges().length, 0, '.targetRanges');
}, 'InputEvent constructor with InputEventInit where targetRanges is empty');

test(function() {
  // Create a StaticRange with arguments (startContainer, startOffset, endContainer, endOffset)
  var startContainer = document.createTextNode("example");
  var endContainer = document.createTextNode("example");
  var staticRange = new StaticRange({
    startContainer: startContainer,
    startOffset: 0,
    endContainer: endContainer,
    endOffset: 1
  });

  var e = new InputEvent('type', { targetRanges: [staticRange] });

  assert_equals(e.getTargetRanges().length, 1, '.targetRanges');
}, 'InputEvent constructor with InputEventInit where targetRanges has one item');

</script>
